Moves PATH to a central location
`zshenv` is the first place to be executed on shell startup. It's
executed by all shells, login or not and interactive or not. But
prezto redirects to the `zprofile` to set up the environment, so the
sourcing of `add_paths` is done there.

This commit also adds the Macbook-specific stuff for the NGV laptop,
which can be removed later, post-gig.